+++ Adapters/LiftoffMonetize/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## LiftoffMonetize iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationLiftoffMonetize', '7.7.3.0'
+pod 'CASMediationLiftoffMonetize', '7.7.4.0'
 ```
+
+### 7.7.4.0
+- Certified with LiftoffMonetize - 7.7.4
 
 ### 7.7.3.0
 - Certified with LiftoffMonetize - 7.7.3