+++ Adapters/YangoAds/CHANGELOG.md
@@ -3,6 +3,9 @@
 pod 'CASMediationYangoAds', '8.1.0.0'
 ```
 
+### 8.1.0.1
+- Fixed minor issues with ad loading.
+
 ### 8.1.0.0
 - Certified with YangoAds - 8.1.0
 - Required minimum AppMetrica version 6.3.0